Added stack trace tracking (DMT_STACK_TRACE)
If the macro DMT_STACK_TRACE is defined, stack information will be
stored in an allocation's node. When dmt_dump() is called the stack
information will be dumped immediately after each node's basic
information. The use of the DMT_STACK_TRACE define is only supported
under Linux, and is not enabled by default.